Apply follow-up bundle review fixes
Address the remaining review findings that still applied in the\ncurrent tree. Add the missing bundle-validator evidence command to the\nexecplan, correct the stale commit_loaded_skill panic text, route\ninstall_skill through install_target_dir, and simplify the skill_fetch\nIP fixtures.\n\nSplit bundle path parsing into a private submodule to keep the module\nunder the file-size guideline, clamp bundle read preallocation against\nthe safe file-size bound, and refresh bundle-test documentation and\ndirectory-rejection wording without changing the underlying validation\ncontract.
