+++ src/agent/dispatcher/tests/skills.rs
@@ -50,7 +50,7 @@ fn install_skill(registry: &Arc<RwLock<SkillRegistry>>, name: &str, skill: Loade
         .write()
         .expect("failed to acquire registry write lock");
     reg.commit_loaded_skill(name, skill)
-        .unwrap_or_else(|e| panic!("failed to commit_install {name}: {e}"));
+        .unwrap_or_else(|e| panic!("failed to commit_loaded_skill {name}: {e}"));
 }
 
 fn make_context_skill(trust: SkillTrust) -> LoadedSkill {
